+++ src/main/java/com/github/noamm9/interfaces/IMonoAudioChannel.java
@@ -1,0 +1,6 @@
+package com.github.noamm9.interfaces;
+
+
+public interface IMonoAudioChannel {
+    void noammaddons$refreshPosition();
+}
