--- src/main/java/com/github/noamm9/interfaces/MonoAudioChannel.java
@@ -1,6 +1,0 @@
-package com.github.noamm9.interfaces;
-
-
-public interface MonoAudioChannel {
-    void monoAudio$refreshPosition();
-}
