+++ src/main/java/com/github/noamm9/mixin/MixinChannel.java
@@ -1,7 +1,7 @@
 package com.github.noamm9.mixin;
 
 import com.github.noamm9.features.impl.misc.sound.MonoAudio;
-import com.github.noamm9.interfaces.MonoAudioChannel;
+import com.github.noamm9.interfaces.IMonoAudioChannel;
 import com.mojang.blaze3d.audio.Channel;
 import net.minecraft.world.phys.Vec3;
 import org.lwjgl.openal.AL10;
@@ -14,7 +14,7 @@ import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 @Mixin(Channel.class)
-public class MixinChannel implements MonoAudioChannel {
+public class MixinChannel implements IMonoAudioChannel {
     @Shadow @Final private int source;
     @Unique private Vec3 monoAudio$lastPosition = new Vec3(0.0, 0.0, 0.0);
     @Unique private boolean monoAudio$relative;
@@ -23,7 +23,7 @@ public class MixinChannel implements MonoAudioChannel {
     private void forceMonoPosition(Vec3 pos, CallbackInfo ci) {
         if (!MonoAudio.INSTANCE.enabled) return;
         monoAudio$lastPosition = pos;
-        monoAudio$refreshPosition();
+        noammaddons$refreshPosition();
         ci.cancel();
     }
 
@@ -32,12 +32,12 @@ public class MixinChannel implements MonoAudioChannel {
         if (!MonoAudio.INSTANCE.enabled) return;
         monoAudio$relative = relative;
         AL10.alSourcei(source, AL10.AL_SOURCE_RELATIVE, AL10.AL_TRUE);
-        monoAudio$refreshPosition();
+        noammaddons$refreshPosition();
         ci.cancel();
     }
 
     @Override
-    public void monoAudio$refreshPosition() {
+    public void noammaddons$refreshPosition() {
         if (!MonoAudio.INSTANCE.enabled) return;
         MonoAudio.applyCenteredPosition(
             source,
