+++ src/main/kotlin/com/github/noamm9/init/types/ICustomMenu.kt
@@ -1,4 +1,4 @@
-package com.github.noamm9.ui.gui
+package com.github.noamm9.init.types
 
 interface ICustomMenu {
     fun isActive(): Boolean
